feat: new project structure (wip) requisitos 1 - 3
