+++ setup.cfg
@@ -1,7 +1,3 @@
 [flake8]
 exclude = .git,__pycache__,.venv
 max-complexity = 5
-
-[tool:pytest]
-addopts = --cov=tech_news_app --cov=tech_news_data_collector
-testpaths = tests/
