+++ tech_news/analyzer/ratings.py
@@ -1,0 +1,6 @@
+def top_5_news():
+    """Seu código deve vir aqui"""
+
+
+def top_5_categories():
+    """Seu código deve vir aqui"""
