+++ tech_news/analyzer/search_engine.py
@@ -1,0 +1,14 @@
+def search_by_title():
+    """Seu código deve vir aqui"""
+
+
+def search_by_date():
+    """Seu código deve vir aqui"""
+
+
+def search_by_source():
+    """Seu código deve vir aqui"""
+
+
+def search_by_category():
+    """Seu código deve vir aqui"""
