+++ tech_news/collector/exporter.py
@@ -1,0 +1,2 @@
+def csv_exporter():
+    """Seu código deve vir aqui"""
