+++ tech_news/collector/importer.py
@@ -1,0 +1,2 @@
+def csv_importer():
+    """Seu código deve vir aqui"""
