+++ tech_news/collector/scrapper.py
@@ -1,0 +1,6 @@
+def fetch_content(url, timeout=3, delay=0.5):
+    """Seu código deve vir aqui"""
+
+
+def scrape(fetcher, pages=1):
+    """Seu código deve vir aqui"""
