+++ tech_news/database.py
@@ -1,0 +1,12 @@
+from pymongo import MongoClient
+
+client = MongoClient()
+db = client.tech_news
+
+
+def insert_or_update(notice):
+    """Seu código deve vir aqui"""
+
+
+def check_duplicates(news):
+    """Seu código deve vir aqui"""
