+++ tech_news/menu.py
@@ -1,0 +1,6 @@
+def collector_menu():
+    """Seu código deve vir aqui"""
+
+
+def analyzer_menu():
+    """Seu código deve vir aqui"""
