--- tech_news_app/news_analyser.py
@@ -1,6 +1,0 @@
-def top_5_news():
-    raise NotImplementedError
-
-
-def top_5_categories():
-    raise NotImplementedError
