--- tech_news_app/news_search_engine.py
@@ -1,14 +1,0 @@
-def search_by_title():
-    raise NotImplementedError
-
-
-def search_by_date():
-    raise NotImplementedError
-
-
-def search_by_source():
-    raise NotImplementedError
-
-
-def search_by_category():
-    raise NotImplementedError
