--- tech_news_data_collector/news_exporter.py
@@ -1,6 +1,0 @@
-def csv_exporter():
-    raise NotImplementedError
-
-
-def json_exporter():
-    raise NotImplementedError
