--- tech_news_data_collector/news_importer.py
@@ -1,6 +1,0 @@
-def csv_importer():
-    raise NotImplementedError
-
-
-def json_importer():
-    raise NotImplementedError
