--- tech_news_data_collector/news_scrapper.py
@@ -1,2 +1,0 @@
-def scrape():
-    raise NotImplementedError
