--- tests/test_news_analyser.py
@@ -1,22 +1,0 @@
-def test_top_5_news_retorna_5_noticias_mais_populares():
-    assert False
-
-
-def test_top_5_news_se_houver_menos_de_5_noticias_retorna_todas():
-    assert False
-
-
-def test_top_5_news_se_nao_houver_noticias_retorna_lista_vazia():
-    assert False
-
-
-def test_top_5_categories_retorna_5_categorias_mais_frequentes():
-    assert False
-
-
-def test_top_5_categories_se_houver_menos_de_5_noticias_retorna_todas():
-    assert False
-
-
-def test_top_5_categories_se_nao_houver_noticias_retorna_lista_vazia():
-    assert False
