--- tests/test_news_exporter.py
@@ -1,34 +1,0 @@
-def test_csv_exporter_extensao_invalida():
-    assert False
-
-
-def test_csv_exporter_cria_arquivo_na_raiz():
-    assert False
-
-
-def test_csv_exporter_substitui_arquivo_existente():
-    assert False
-
-
-def test_csv_exporter_cabecalho_invalido():
-    assert False
-
-
-def test_csv_exporter_sucesso():
-    assert False
-
-
-def test_json_exporter_extensao_invalida():
-    assert False
-
-
-def test_json_exporter_cria_arquivo_na_raiz():
-    assert False
-
-
-def test_json_exporter_substitui_arquivo_existente():
-    assert False
-
-
-def test_json_exporter_sucesso():
-    assert False
