--- tests/test_news_scrapper.py
@@ -1,22 +1,0 @@
-def test_scrape_raspa_primeira_pagina_por_padrao():
-    assert False
-
-
-def test_scrape_raspa_n_paginas_se_definido():
-    assert False
-
-
-def test_scrape_trata_404():
-    assert False
-
-
-def test_scrape_atributos_obrigatorios():
-    assert False
-
-
-def test_scrape_atualiza_noticias_repetidas():
-    assert False
-
-
-def test_scrape_exibe_mensagem_apos_finalizar():
-    assert False
