--- tests/test_news_search_engine.py
@@ -1,34 +1,0 @@
-def test_search_by_title_se_encontrar_retorna_a_lista_de_noticias():
-    assert False
-
-
-def test_search_by_title_se_nao_encontrar_retorna_lista_vazia():
-    assert False
-
-
-def test_search_by_date_se_encontrar_retorna_a_lista_de_noticias():
-    assert False
-
-
-def test_search_by_date_data_invalida():
-    assert False
-
-
-def test_search_by_date_se_nao_encontrar_retorna_lista_vazia():
-    assert False
-
-
-def test_search_by_source_se_encontrar_retorna_a_lista_de_noticias():
-    assert False
-
-
-def test_search_by_source_se_nao_encontrar_retorna_lista_vazia():
-    assert False
-
-
-def test_search_by_category_se_encontrar_retorna_a_lista_de_noticias():
-    assert False
-
-
-def test_search_by_category_se_nao_encontrar_retorna_lista_vazia():
-    assert False
